update test case
