+++ featuretools/tests/computational_backend/test_calculate_feature_matrix.py
@@ -314,6 +314,38 @@ def test_training_window(es):
     assert (feature_matrix[dagg.get_name()] == dagg_values).values.all()
 
 
+def test_training_window_overlap(es):
+    es.add_last_time_indexes()
+
+    count_log = ft.Feature(
+        base=es['log']['id'],
+        parent_entity=es['customers'],
+        primitive=Count,
+    )
+
+    cutoff_time = pd.DataFrame({
+        'id': [0, 0],
+        'time': ['2011-04-09 10:30:00', '2011-04-09 10:40:00'],
+    }).astype({'time': 'datetime64[ns]'})
+
+    actual = ft.calculate_feature_matrix(
+        features=[count_log],
+        entityset=es,
+        cutoff_time=cutoff_time,
+        cutoff_time_in_index=True,
+        training_window='10 minutes',
+    )
+
+    expected = pd.DataFrame({
+        'COUNT(log)': {
+            (0, pd.Timestamp('2011-04-09 10:30:00')): 1,
+            (0, pd.Timestamp('2011-04-09 10:40:00')): 9,
+        },
+    })
+
+    assert actual.equals(expected)
+
+
 def test_training_window_recent_time_index(es):
     # customer with no sessions
     row = {
